Bug#526392: [PATCH] Provide default values for tg2quilt.mk variables
tg2quilt.mk currently requires that $(QUILT_PATCH_DIR) and
$(QUILT_STAMPFN) be set explicitly, either via quilt.make or manually.

This makes $(QUILT_PATCH_DIR) default to debian/patches, which is the
location used by most packages, now made semi-official with the 3.0
(quilt) source format.

As for $(QUILT_STAMPFN), it is set to "patch" by default, which is
probably the safest and blandest value, being the recommended target
name per Policy.

Signed-off-by: martin f. krafft <madduck@debian.org>